MBS-12942: Also indicate if target has pending edits in rels editor
We were already showing if the relationship itself has pending edits,
but not if the target entity has them (which the old relationship editor
used to indicate with a highlight).
Aerozol made us a nice icon to use for this, and I think
it's quite useful to have it and link to the entity's open edits page.