MBS-13001: Actually set source_type and target_type for historic rels (#2900)
These being missing was causing a beta ISE when combined with
commit 268d6e51db045be90b3b2ec49d28190d2cfb1fe5 -
which makes us a lot more demanding wrt what we send to linkedEntities.

Basically, the TO_JSON method for Entity::Relationship adds the rel source
to linkedEntities using the source ID, the source itself,
but for some reason the source_type for the relationship entity
rather than source->entity_type.
This might be in case the "source entity" is
a historical Frankenstein monster of an object, but in any case
the historical relationships were not setting a source_type at all,
and unsurprisingly "" is not a valid type assignable to linkedEntities.

For what it's worth, _display_relationships seems to already
add the source to linkedEntities itself, which is probably why these
edits still worked fine. We might want to drop the double linking?