MBS-11393: Show SetTrackLengths edit in recording edit history (#1969)
If a medium edit changes the duration of a recording,
it is shown on the recording's edit history. If a discID does
(with a set track lengths edit), it does not, so it becomes
hard to see where the set duration came from.
This will add the set track lenghts edit to the history
of all recordings which are linked to a track that is changing
duration.