MBS-12990: Don't pad the legend relationship editor fieldset
These legends are all supposed to be aligned, but the padding
being applied to the whole fieldset was moving this one further right.
I was planning to just add padding to the table inside,
but I guess it would also make sense to keep errors padded,
so I added an extra div for it.