Replace CoreEntityT with RelatableEntityT in EntityLink
Same reason as the change in entityHref:

"This makes sense to me because anything that's relatable must be able to be
linked to via display of the relationship itself."

These types are currently equivalent in Flow.  I'm not asserting they've always
been equivalent, just that any non-relatable types that are linkable in the
future can simply be added to the existing union.

Co-authored-by: yvanzo <yvanzo@metabrainz.org>