MBS-12800: Exclude cancelled releases from first_release_date
A release marked as cancelled by definition does not have a release date
and as such should not be used when calculating the first release date
of recordings or release groups connected to it.

This amends the get_release_first_release_date_rows function
to ignore releases with status 6 (cancelled).

It also updates the a_upd_release function to recalculate first dates
when a status is changed to or from 6 (cancelled).