MBS-13007: Fix a regressive ISE on broken WikiLink
When a wiki page contains a WikiLink to a page that doesn’t exist yet,
such as `Category:Relationship_Type` in the page `Terminology`,
the class `new` is set for this link which contains text only.

I overlooked this case in commit 45288c8b merged with pull request #2894
to address MBS-12984, leading to ISE on `tag` method not found.

This patch fixes that by checking that the content is an element first.