MBS-12800: Update materialized tables during upgrade
The previous commit changed the relevant function so that
cancelled releases won't be taken into consideration in the future.
The upgrade script does need to also deal with cancelled release data
which has already affected the materialized tables, though.
The simplest way to do this seems to be to just drop the relevant rows
from the release table, and then recalculate the recording and RG data
for any recordings and RGs related to cancelled releases
(this should update the data based on the new release table
which no longer has the cancelled release data).