MBS-12794: Update calculated tags and ratings when user set to spammer
When we mark a user as spammer, we won't delete their tags and ratings,
since they can be useful for spam fighting machine learning.
We do not want to have them count when calculating
the aggregate tags and ratings for entities, though.
As such, this trigger should remove them from the general calculations
when the user is set as spammer, and bring them back
to those calculations if the spammer flag is unset.