MBS-13025: Escape link symbols in setlist syntax
Having square bracket `[` or `]` on the same line as a link
`[mbid|name]` currently bugs its rendering in an event setlist.

This patch allows to escape these two symbols using HTML entities.
It also allows to escape the ampersand `&` which starts those.
HTML entities are quite well-known and supported in Wiki and Markdown.

It recommends using `&lsqb;`, `&rsqb;`, and `&amp;` HTML entities, but
it also supports all equivalent HTML entities which will be documented
in https://musicbrainz.org/doc/Event/Setlist and which are taken from
https://en.wikipedia.org/wiki/List_of_XML_and_HTML_character_entity_references