Build Redis connection lazily
This allows running LoadReplicationChanges without Redis installed or running.

`server` and `database` are added as attributes so that we can access them from
`_build_connection` later. These are already passed in as args to the
constructor. We can also remove `sub BUILD` entirely, because it no longer
serves any purpose (each of the arguments to the constructor now match their
attribute's name).