LoadReplicationChanges: Don't error after applying an aborted sequence
If the LoadReplicationChanges process is aborted and replication is later
resumed, it will attempt to continue applying any existing data that was loaded
into the dbmirror pending tables first before downloading the next packet (i.e.
"[c]ontinuing a previously aborted load").

As a reminder, the changes inside the packet are grouped by transaction ID, so
ProcessReplicationChanges applies each group in a separate transaction (as they
were on the master). The corresponding pending data entries are deleted in the
same transaction that they're applied in, so they're only removed once they're
applied (i.e. a transaction cannot be half-applied if the script is aborted).

However, the script would always error after applying the leftover changes due
to `$REPLICATION_SEQUENCE` being undefined: it's only read from the packet, and
we don't have any access to the previous packet data when handling the aborted
load. This commit skips the sequence comparison in that situation.