ProcessReplicationChanges: Use repeatable read transaction isolation level
As noted in 04c3c47f08e70135f07be66b44324cf75dc12529, the serializable
transaction isolation level isn't needed here (and wasn't working to begin
with), but I turned it on because it matched the intent of the code and didn't
seem to hurt. But it does cause at least one issue; this error has occurred
ocasionally while running slave.sh on aretha:

53200 DBD::Pg::st execute failed: ERROR:  out of shared memory HINT:  You might
need to increase max_pred_locks_per_transaction. [for Statement "FETCH FORWARD
10000 FROM csr"] at
/home/bitmap/musicbrainz-server/admin/replication/../../lib/Sql.pm line 94.

The PostgreSQL documentation says that "you might need to raise
[max_pred_locks_per_transaction] if you have clients that touch many different
tables in a single serializable transaction."

Instead, I'm changing the transaction isolation level to REPEATABLE READ, which
is not as strict as SERIALIZABLE, but still more strict than the default level
used prior to 04c3c47f08e70135f07be66b44324cf75dc12529. And as explained in
that commit, there should only be a single transaction modifying data on
mirrors to begin with (from the hourly ProcessReplicationChanges script). So
serialization anomalies [1] aren't really a concern.

[1] https://www.postgresql.org/docs/12/transaction-iso.html