MBS-13042: Increase default font size (12 -> 14px)
A fair amount of users (including me) seem to always use MBS zoomed in
to at least 110% just because of how small the default font is.
Apparently 16px is the recommended accesibility default,
but given our current design might not accomodate it too well,
this changes the default from 12px to 14px for now, which should already
be fairly noticeable.
I did some basic testing around the site and found it perfectly usable
without any weird wrapping, but this would require proper user testing
by more than just me, so we should put it in test and then beta
and collect feedback.

As a note, rem values in here are based on HTML font size,
which is @default-text-size (set in layout.less).

Co-authored-by: santiago fn <santiagofn@users.noreply.github.com>