MBS-13052: Support intl-XX/ path segment in Spotify URLs
Apparently some languages are now adding a segment like intl-de
or intl-it to the Spotify URLs. These redirect to the basic URL
if the site is set to English, and the basic URL to them if not.
As such, the easiest way to deal with them seems to be
to just drop them in cleanup.