MBS-13060: Display attribute MBIDs
We have MBIDs in the DB for all these attributes,
but we do not  currently display them anywhere on the website
(we do use them in the API).

These Attribute pages are also going to be made public
(without the admin bits) at some point with MBS-11178,
and then it would make sense to make it so that mbid/$attribute_mbid
would redirect to the appropriate attributes page (MBS-13067).