MBS-7385: Use EntityLink over release name in medium append UI
We were just printing "was medium 1 on release Whatever"
on the UI for picking medium positions and titles when merging
releases and appending mediums.
This means release disambiguations were not shown here.
This patch changes the text name to an EntityLink for the release,
which both shows disambiguations and provides
an actual link to the release itself
in case the user wants to reopen it to check something.