MBS-13058: 404 /applications/revoke-access if token does not exist
Right now you can enter whatever integers in the revoke-access URL
and it will try to revoke your token for that application and scope.
Obviously, that's unlikely to exist, and PSQL fails, causing an ISE.

Given we already have the check_granted_token method
which takes the same parameters, we can check
whether there's something to delete before we render the form.
I chose a basic 404 error with custom message rather than
a NotFound page because the only way a user should get here anyway is
if they edit the URL by hand or reopen a tab revoking
an already revoked token.