MBS-13057: Show info about application to be revoked in /revoke-access
Right now once the user clicks through to the revoke-access page,
there's no longer any info printed about what access they are revoking.
The user is expected to click blindly to revoke access to whatever it is
and might end up revoking the wrong thing if they misclicked
in the applications page.

This just lists the application name and its permissions
when asking for confirmation for the revoking.