Block changing edit notes when isAddingNotesDisabled
It seems sensible that if someone has been blocked from adding notes
they should also be blocked from modifying them.
Deleting the note would seem safer, but you can leave a deletion reason
so there's still a way for people to troll through it,
and it's cleaner and simpler to block both options.