MBS-13075: Indicate the number of items in a series on search results
It seems useful to be able to see at a glance the amount
of entities in each series in a search result.

This requires loading the entity_type for the series type
since that is missing from the indexed search data. That in return
needs the type gid, which we have in the search response.
For some reason, we were creating a type with just the name
in schema_fixup_type; this adds the type gid to the fixed up types,
and then loads the entity type for the relevant series types
with load_entity_type_from_gid.