Validate modified edit note like new one
When modifying an edit note, rather than simply requiring the text field,
this actually validates that the edit note is valid and non-empty
and shows appropriate error messages.
Unlike when adding a note, this is not limited to cases where a note
is required, since if the user wants to get rid of the note,
they should just be deleting it instead.

We skip running the validation when the user clicks Cancel,
since we obviously don't care about the quality of the note
they are *not* submitting.