LoadReplicationChanges: Exit code for schema sequence mismatch
After LoadReplicationChanges reaches the first packet of the next schema
sequence, it outputs "You must upgrade your database in order to apply this
replication packet" and exits with exit code 1. This in turn causes the
incremental JSON dump scripts to fail, with admin/RunIncrementalJSONDump not
rsyncing any dumps created prior to the failure.

To avoid this, I'm using a different exit code (3) to indicate a schema
sequence mismatch, and then ignoring this code in JSONDump::Incremental.