MBS-13085: Clarify "share email" report checkbox
While in the user contact form this checkbox really does
"reveal my email address" (since the recipient has no way to see it),
on the report a user form it does not.
An admin can always see your email address - what this really is
is a way of indicating whether you want
to be notified about the report further or not.

Co-authored-by: Aerozol <aerozol1@gmail.com>