MBS-13047: Do not assume every wiki image has /thumb (#2923)
This was assuming /thumb would be in all wiki image URLs,
but at least the ones in Release/Packaging did not have it.
It might be an SVG only thing? In any case, given we want to link
to the full image, we can just keep removing it if there
and ignore it otherwise AFAICT. Similarly, the NN-px bit
seems to be only relevant to thumbs, so we should only remove it
when /thumb is present to avoid breaking non-thumb images
which happen to start with that pattern.