Avoid area bubble crashing on missing iso_3166_1_codes
Using {this} rather than {this.toJSON()} meant we were hitting
the issue why we originally added the iso_3166_1_codes assignment
to toJSON in the first place, and the area bubble was crashing.
In order to be able to display containment with toJSON
we need to actively set it in toJSON as well.

@mwiencek suggested adding iso_3166_2_codes and iso_3166_3_codes
as well for safety.