Add renderToStaticMarkup wrapper
formatSetlist was adding a non-expected comment in the result
because ReactDOMServer.renderToStaticMarkup adds it for some reason.
This gets worse though - when using the original renderToStaticMarkup
multiple times (for a future test),
the comment only gets added to the first of them.
As such, it would have meant removing the comment from formatSetlist
and adding it to the new test file instead,
and moving it around each time we use renderToStaticMarkup again
in a new test (at least one that sorts alphabetically earlier...).

That's just ridiculous IMO, so adding a wrapper that just returns
the result of the ReactDOMServer function, but strips
any initial empty HTML comments.