Use compareDates as the base for areDatesEqual
areDatesEqual was actually failing when one date part
was explicitly set as null in one date but ommited in the other.
That does not seem intentional, given we would consider those two
the same date in compareDates.
More generally, there seems to be no reason to have two implementations
of a way to check whether two dates are the same.
This keeps areDatesEqual for clarity of function naming,
but simply by making sure compareDates thinks
the two dates are the same.