MBS-13104: Unblock and handle RYM interview links
We were blocking these, but they seem to be perfectly legit.

I wasn't sure whether to auto-select only based on the
"sonemic-interview-" start of the URL subsection, but I'm not confident
enough that it will never change. In any case,
we only allow /feature links for interviews now.
If we change this in the future we can be more strict with validation.