MBS-13112: Hide username field when changing password
While I understand this field exists for the use case where the user
is not logged in and the username value is passed as a string parameter,
it doesn't make any sense to have this option for the default use case
when a logged in user wants to change their own password.
In that case, the user is already under their user header
and having the option to select a different username here
is very confusing.