MBS-13111: Hide list of autoeditor election voters during vote
This being shown during the voting period has no actual benefit,
and the drawback of allowing people who can see the vote tally
(proposer, seconders and candidate) to figure out who voted what
by comparing the vote tally with the vote list repeatedly upon changes.

The only real usefulness I can see in showing the voter list
during the election period is that it allows voters to see
what they have voted in case they want to change it.

As such, this commit hides the voters list during the election period
and indicates the current vote if any of the page visitor with a line
above the voting buttons.