MBS-13116: Also list abstaining votes in vote tally
In the same way we let the proposer, candidate and seconders
know the amount of Yes and No votes while the election is happening,
it can be useful to know the number of abstaining editors as well
(if it's a lot, or even a majority, then that suggests the election
might have been premature or the introduction of the editor
in the autoeditor forum not good enough).

This adds it to the tally. The alternative option
(showing number of votes cast) also seems sensible so adding that too,
but there seems to be no reason why we'd make people do math for this.