Remove Entity::CoreEntity
We decided to get rid of CoreEntity, and all CoreEntity users
are also Relatable, so merging it into Relatable and just extending
Entity with that.

Doing this breaks the sub name in Entity::URL, (seems it gets
overriden by Role::Name) so changing it to has 'url' with a default,
which still works.