MBS-13120: Fix note-too-old check for modifications
According to
https://metacpan.org/pod/DateTime#$dt-%3Esubtract_datetime($datetime) (which
from the source code appears to be what `-` is overloaded to), the returned
duration "may have deltas for months, days, minutes, seconds, and nanoseconds."
The key word being "may."  When I reproduced the test failure in
`Controller::Edit::ChangeNote` locally, the duration object had values for
months, minutes, and seconds, but not days.  This may vary depending on what
time the test is run.

Adding a duration of 1 day to the post time and checking if it's less than now
fixes the test for me locally. I added an additional test to make sure we're
not preventing changing an edit note posted at 23:59:59 after 1 second (i.e.
if the day changed, but the overall duration is still less than a day).