MBS-12026: Still change deleted editor votes if they lack privs
We were being so respectful of the "Editing and voting disabled" flag
that we were following it even when the editor was deleted
and we wanted their votes to be changed to Abstain.
This allows passing an $override_privs boolean to enter_votes
which will skip checking whether the editor has voting rights.
In order to pass that I needed to stop passing @votes as an array
and pass an array ref instead, hence the other changes in this commit.