admin/run: fix log path and automatically create it
sitemaps cron has not been running for quite some time.  I'm not sure for how
long, but at some point the logs volume was cleared, which breaks admin/run.

To check what was failing, I ran this inside the container:

```sh
root@0ca0886613a0:/home/musicbrainz/musicbrainz-server# ./admin/run admin/cron/daily-sitemaps.sh
run: '/root/log/daily-sitemaps' doesn't exist
Make sure to run in the appropriate container.
```

This is odd to me because the script is supposed to `sudo` as the musicbrainz
user if it exists.  In that case it should also use /home/musicbrainz as the
log path, since that's the user it's invoking the passed-in script as.

Next, I tried the same command as the musicbrainz user:

```sh
root@0ca0886613a0:/home/musicbrainz/musicbrainz-server# su musicbrainz
musicbrainz@0ca0886613a0:~/musicbrainz-server$ ./admin/run admin/cron/daily-sitemaps.sh
run: '/home/musicbrainz/log/daily-sitemaps' doesn't exist
Make sure to run in the appropriate container.
```

Since the log volume was cleared at some point, these directories don't exist,
and the script fails.  This was never an issue before, because our old crontab
automatically created the log path:

```
-10 0 * * * . /etc/container_environment.sh; mkdir -p $HOME/log/daily-sitemaps && cd $HOME/musicbrainz-server && carton exec -- admin/cron/daily-sitemaps.sh > $HOME/log/daily-sitemaps/$(date --utc +\%FT\%TZ).log 2>&1
```

It appears that the intent of the script is to error in case you're running the
command inside the wrong container, but I think that rarely happens in
practice.  What is clear is that we don't actively monitor the sitemaps cron
logs for failures, so I'd much prefer the script be made more resilient.

I've made the following changes to resolve these issues:

 * Storing the username we're invoking the command as in `$username` and
   putting logs in `~$username`.

 * Creating `$log_dir_path` instead of erroring.

 * Ensuring the created log path and file are owned by `$username`.