MBS-13094: Allow admins to browse edit note changes
Admins should be able to see the edit history for any edit,
since otherwise a user could, for example, insult another
and then edit the note so that the email is sent but others don't see.

This mostly follows what we do for annotations, but most of it does not
support translated strings since it's admin-only.
Translations are added for stuff that will show
in otherwise translated pages.

Since the admin could theoretically try and visit a /changes page
for a non-existing edit note, I made the code return an error
in that case rather than an ISE.