MBS-13092: Support Apple Music song links
We were blocking song links, but I can't see any real reasons to.
There's two kinds of song links - one that selects the track on an album
(kind of like our track links in MB) and one that goes to a /song page
for the track itself. There's no reason not to convert the former
into the latter, as far as I can tell.