Drop decoda.com from malware domain list
decoda.com (an old lyrics page that became dodgy
and we blocked as dangerous) is now the home page for a company
that produces mine site technology. Obviously we shouldn't link to it,
but it isn't malware anymore and we don't need to specialcase it either.
yvanzo checked that "it’s not listed in URL reputation databases either"