Fix release search log flood since commit e38d8e9 (#2978)
Since the commit e38d8e95e6ab06888c956df668a4cb5617a3b4ea is in prod
(merged with the pull request #2902 and released with v-2023-06-07),
release search queries are logging two warnings for each result, that is
up to 200 log lines for each query, resulting in needless GBs of logs.

Search is indeed returning releases having mediums with searchable fields
only (`format` and `track_count`) so without `position` and `release_id`.

This patch simply checks if defined before enforcing integer conversion,
that is the same approach for `position` and `release_id` as for
`format_id` and `track_count`, thus stopping the log flood.

However by doing so, it partly reverts the commit e38d8e9 as it unsyncs
the Flow type `MediumT` for its `position` and `release_id` properties.
It is unclear whether these should be nullable or if a new type should
be used instead since these are null only for search results. Beyond
Flow types, the property `mediums_loaded` in Perl is misleading too.
A larger refactoring might be needed to sanitize the code.