MBS-13160: Rename "legal name" on artist page to "performance name of"
We renamed the relationship text for "performs as" with STYLE-2328
but the display on the artist overview page was still hardcoded
to "Legal name". This renames it to "Performance name of"
(as in the relationship) in the case where we follow this relationship
(which no longer mandates it must be a legal name).
"Legal name" is kept for actual legal name alias display.

I wasn't sure what'd be a good name for the no-longer-legalName,
ended up going with baseName since it's the base of the grouping
of names for the artist.