Switch Babel & ESLint parsers to hermes
This is officially recommended by
https://flow.org/en/docs/tools/babel/
and
https://flow.org/en/docs/tools/eslint/
in order to support the latest Flow syntax (type guards, mapped types,
conditional types, and I believe soon a new component syntax) without
having to wait for these to be added to Babel.

eslint-plugin-flowtype doesn't work with this parser, but there is a
community-maintained clone called eslint-plugin-ft-flow which does.
Just swapping the plugins doesn't work without a few changes:

 * The no-undef rule is broken. It isn't essential since Flow
   checks that for us, so I've disabled it.

 * The no-unused-vars rule is triggered for all global Flow types,
   so I disabled it in all files under root/types.