Upgrade Flow to 0.208.0
https://github.com/facebook/flow/releases/tag/v0.208.0

The following change has caused some new errors:

"Flow will now error on unsupported statements in library definitions, instead
of silently ignoring them."

This affected root/vars.js, where we have some `function` definitions. These
are not allowed in libdefs, apparently, or at least not anymore. That's an
issue for us: we rely on `%checks` in order to refine the input types, and you
can't otherwise use `%checks` in a function type (it requires a definition!).

(Interestingly, the fact that `%checks` was working completely as intended
prior to this Flow version contradicts their claim that it was silently ignored
in the libdef before.)

We're partially in luck, because there's a new "type guards" feature that can
be used to refine function inputs in libdefs. However, it's experimental and
appears to be buggy -- while I've made use of it here, I had to add a lot of
`$FlowIssue` comments too. I believe the bugginess with type guards is fixed in
later Flow versions, so I'll be able to remove these comments in a subsequent
commit.