MBS-13162: Handle Threads links (#2981)
This adds threads.net links to the sidebar, and cleans them
to use https://www. which seems to be the default.
It also selects (mandatory) social media for any entity types
that support it (currently artist, event, label, place and series),
while selecting stream for free for recording and blocking other
free streaming entities.
Social media links must be user pages, while streaming links
must be threads ("tweets").

This matches what we do with Twitter, which this is mostly
a straight clone of.