MBS-13173: Also allow daily.bandcamp.com interviews
Bandcamp publishes interviews with artists seemingly quite often,
but we were automatically selecting "bandcamp" for those.
This moves daily.bandcamp URLs to a separate handling section,
so we can deal with them more properly and select interview
when appropriate.

While doing this I noticed that newer album reviews are no longer listed
under the date as we expected, but under an album-of-the-day path.
As such, I made both interview and review support these sorts of paths.

Finally, since we are no longer doing any RG-level selecting
on the standard bandcamp handling, non-daily bandcamp URLs
need to be blocked separately on RGs.
AFAICT no non-daily bandcamp URL would be acceptable for RG
(and the previous setup blocked them all)
so this is a blanket ban.