MBS-13177: Show label events on label rels page
Event-Label rels were already being loaded (unpaged),
but not being shown since Relationships.js did not use them.
This was not a problem until recently because we didn't have any.

The easiest fix would be to just use them in Relationships.js.
Other option would be to add an events tab for labels like we have
for artists.

Personally, neither option seems great. A tab seems too much given
that events are a lot less important to (most) labels than they
are to (most) artists, and just listing them in the relationships
section seems like it could be overwhelming for the sort of
organizations, like a concert promoter, that specifically
get related to a lot of events. Also, the only Event-Label we have
so far is label presents event, a 1:0 cardinality relationship,
and we usually show these sorts of relationships in a table,
which we already do for other things like "copyright on".

Given all this, I decided to load Event-Label rels as a paged subset
and show them in RelationshipsTable.js instead.