MBS-13209: Add validation for Juno Download URLs
There's three sets of pages we probably want to link to:
/artists, /labels, /products (releases).

There's also genre pages but we don't have a good way to link them
to genres right now so this does not do anything with them.

There's three other kind of pages we had existing links to:

/plus pages no longer exist and it seems they were a blog/review site,
which is fine to no longer support since it is gone.

/search pages are blocked with the usual "don't link to search" message.

/charts seems to be DJ-curated playlists that can be purchased as a block.
It's unclear whether these are equivalent to a MusicBrainz release, but
we only have one link to these in 20k Juno links, so it seems fine
to just not handle them (block them) for now and have a discussion on
how they map to MB if and when someone requests them.