MBS-13208: Clean up Juno Download URLs further
When adding validation it became clear that we probably want to clean
the urls further to avoid, say, two links for tracks and releases
for the same artist. This does that.
Artists and labels only have the name after the prefix,
while releases have text plus a numeric ID in most cases, but
just the id is also valid, so we clean up to that,
as per the test examples.