MBS-13212: Handle Behance links
These should use the new "art gallery" relationship for artists.
User pages are just behance.net/user so I clean away
everything else - that same structure matches a bunch of hardcoded
paths that I'm also blocking here just in case.

We have only 200 or so of these so not adding them to the sidebar yet.