MBS-10977: Handle Dribbble links
These should use the new "art gallery" relationship for artists.
User pages are just dribbble.com/user so I clean away
everything else - that same structure matches a bunch of hardcoded
paths that I'm also blocking here just in case.

We have only 30 or so of these so not adding them to the sidebar yet.