MBS-13216: Add VNDB to other databases whitelist
This also adds a basic cleanup and validation.
As far as I can tell, they have pages for "staff" (/s, artist),
"character" (/c, artist), "producer" (/p, label) and "visual novel"
(/v, which would probably be works, but we don't actually use works
for games right now and this is not a soundtrack page, so I'm
blocking for now).