MBS-13218: Add Jaxsta to other databases whitelist
This also adds a basic cleanup and validation.
As far as I can tell, they have pages for "profile" (artist and label),
"recording" and "work" (which map to ours), and "release".

"Release" is a double-UUID link, where the first UUID is basically
equivalent to our release group, and the second is a specific
variant of the group (a MB release). In an ideal world we would
remove the second UUID when trying to link from a release group,
and keep both when linking from a release. We do not have support
for anything like this when cleaning URLs though, so for now we just
map the variant to releases. The site automatically appends a variant
when loading a single-UUID release link anyway, so users should not
be confused here.