MBS-13218: Allow linking Jaxsta works as lyrics
Jaxsta recording, release (for singles) and work pages
can contain (licensed) lyrics but that's not their primary function,
so using `otherdatabases` by default still seems preferable;
for works this also allows manually selecting lyrics in addition,
as we do for Mainly Norfolk. For releases that is restricted
to singles and for recordings we don't have an equivalent
lyrics option anyway, so not allowing lyrics there.