MBS-13222: Add MobyGames to other databases whitelist
This also adds a basic cleanup and validation.
As far as I can tell, they have pages for "person" (artist)
and "company" (label) that are a good mapping to MB now.
They also have pages for specific games that could be linked to RGs
(for cases where we have releases that represent actual game discs)
but they seem too likely to be misused for the far more common situation
of wanting to link the soundtrack album to the game page,
which we don't currently support as part of a standard otherdbs rel.
Same for work soundtracks <-> game pages.
As such, this blocks game links for the moment.