Amend 717d2a05b5: Remind that Perl $! is $OS_ERROR
It is the same Perl variable just with a shorter name.
https://perldoc.perl.org/perlvar#$OS_ERROR

Short names are not allowed by the Perl linter.
But it is not checking such Perl snippets embedded in Bash scripts.
Using `$OS_ERROR` requires `use English`, so just commenting with the
meaning of this Perl variable as the Bash `$!` has a different meaning.