MBS-13207: Make release recording input not steal focus (#3008)
Prevent the focus from getting incorrectly reassigned to the
recording name text field whenever a radio button is
selected at /release/add#recordings.

The RecordingBubble was passing a stealFocus argument around
internally but then ignoring it in its show() method and not
passing it to its parent class BubbleDoc (which also ignored
it). Make sure that the argument makes it all the way up to
the BubbleBase class.