MBS-13234: Handle international Ticketmaster URLs
Turns out there's a ton more international Ticketmaster sites
that we weren't handling with the previous setup.
See https://www.livenationentertainment.com/global-sites/
They seem to be similar to the parent US one, but they have
per-country domains (either one level, such as .lat or .sg,
or multilevel, such as co.nz or co.za),
and the name slugs are sometimes in different positions in the URL.
See tests for examples.

On the sidebar this just shows Ticketmaster COUNTRY DOMAIN
which is probably clear enough and similar to what we do
with Amazon and iTunes.