MBS-2604: Infer link direction between people and groups (#3011)
Make the relationship editor automatically infer the link
direction for certain artist-artist relationships between
individuals and groups (including orchestras and choirs).

For example, when a "member of" relationship with a person
target is added to a group artist, it is automatically
reversed.

This is intended to reduce the need to click the "Change
direction" button and to make errors less common. Editors
are still free to manually change the link direction before
saving the relationship.