MBS-13241: Prevent recording bubble from shifting.
Set "vertical-align: top" on the tracklist table rows at
/release/add#recordings to prevent the "Edit" button from
shifting vertically depending on the height of the track and
recording titles, which was causing the bubble that contains
the recordings list also shift vertically.

Also set "vertical-align: top" on the recording rows since
it looks odd when the artist name and duration aren't
aligned with the top of a multiline recording title and
disambiguation.