MBS-13260: Upgrade Node.js to v18
From the ticket: "MusicBrainz Server depends on Node.js 16 which is currently
still an LTS in maintenance phase. Its EOL is the 11th September 2023. So it
should be replaced with the next active LTS which is Node.js 18."

Since we plan to require v20 when that reaches LTS, and given the short notice,
I haven't updated the requirements in INSTALL.md yet (though I've noted that we
run v18 in production); I've also modified the CircleCI config to run tests
against the previous musicbrainz-tests image which contains Node.js v16 in
order to avoid any breakage in the meantime.