MBS-10629: Add cleanup and validation for Anghami URLs
Anghami is an Arab-world-centric streaming service for which
we have around 1500 links already, so it makes sense
to support it properly.

The service has a free, ad-supported tier, so it seems like a good fit
for the "stream for free" relationship.

There are four main types of links we should support:
/artist, /album (release), and /song and /video (recording).