MBS-13224: Show edit/note restrictions from user profile
We have for a long time chosen not to show whether a user was
blocked from editing or leaving edit notes publicly in order
to avoid publicly shaming editors.
Sadly, this means already banned editors get reported again, kind of
wasting the time of both re-reporters and admins.
Additionally, it's unfair to the user *not* to display this info,
since other editors can demand the user react to their edit notes
without knowing they cannot.

This shows active editing/voting an edit note restrictions for the user
to other logged in users only and only when they visit their profile,
which is the least shaming way of doing it but still visible
if someone is thinking of re-reporting.

The "Untrusted" restriction is explicitly not shown to other users.
"Untrusted" is usually set as a "hidden" "minor" restriction
for users who just need an extra eye on their edits,
and we don't even notify the restricted user of it,
so it wouldn't make sense to notify others.