Fix nonsensical "collections tab" test
This was testing for a feature we do not provide, which is hiding
the collections tab if collections are private:
public/private is a parameter of each separate collection.

This might have made sense back in the day (I forget),
since pretty much the same expected result here is actually true
for tabs we *do* allow to make fully private.
As such, making this test the Tags tab (and also actually making sure
it does not display the tab when private, which the test did not check).

Just checking for the tab title isn't good, since the word is
on the menu when logged in, so this checks for the explicit link
and tab title combination.